app/requirements.txt
aiofiles==22.1.0
aiosmtpd==1.4.2
amqp==5.1.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
atpublic==3.1.1
attrs==22.1.0
bcrypt==4.0.0
billiard==3.6.4.0
boto3==1.26.16
botocore==1.29.16
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.68.2
google-api-core==2.10.2
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.21.0
httptools==0.2.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
jmespath==1.0.1
kombu==5.2.4
oauthlib==3.2.2
packaging==21.3
passlib==1.7.4
Pillow==9.3.0
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==4.21.9
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.14.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.15.0
uvloop==0.17.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.4
wrapt==1.14.1
zipp==3.8.1
